% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{convert_am_pm}
\alias{convert_am_pm}
\title{\preformatted{  Convert between AM and PM variablities
}}
\usage{
convert_am_pm(x,
  pm_regex = "(p\\\\.m\\\\.|P\\\\.M\\\\.|p\\\\.M\\\\.|P\\\\.m\\\\.|pm)",
  am_regex = "(a\\\\.m\\\\.|A\\\\.M\\\\.|a\\\\.M\\\\.|A\\\\.m\\\\.|am)")
}
\value{
\preformatted{ AM or PM if matched, x if not
}
}
\description{
given an AM/PM string, convert to the appropriate string for
}
