% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{convert_date}
\alias{convert_date}
\title{\preformatted{  convert date to POSIXct datetime
}}
\usage{
convert_date(date, format, tz = "UTC")
}
\value{
\preformatted{ a vector of POSIXct dates
}
}
\description{
given a date and format, convert
}
