% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{convert_month_language}
\alias{convert_month_language}
\title{\preformatted{  retrieve the timzone lookup table from wikipedia
}}
\usage{
convert_month_language(x, lang = "spanish")
}
\value{
\preformatted{ a data.table ready for insertion into a database or other use
}
}
\description{
this function is used to create the dataset to lookup
}
