% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{convert_word_to_num}
\alias{convert_word_to_num}
\title{\preformatted{  convert a word to a number
}}
\usage{
convert_word_to_num(word)
}
\value{
\preformatted{ a list of the word and the output
}
}
\description{
provided a num ber word convert it to a number
}
