% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{extract_county_number}
\alias{extract_county_number}
\title{\preformatted{  extract a scounty number
}}
\usage{
extract_county_number(x)
}
\value{
\preformatted{ a vector in the form of county/cases
}
}
\description{
extract a county and number and return as a vector
}
