% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{extract_county_number_to_df}
\alias{extract_county_number_to_df}
\title{\preformatted{  extract the county name and county number from string
}}
\usage{
extract_county_number_to_df(x)
}
\value{
\preformatted{ a data.frame with county and cases field
}
}
\description{
given a string that contains county name and number
}
