% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{fetch_number_in_text}
\alias{fetch_number_in_text}
\title{\preformatted{  retrieve a number from text
}}
\usage{
fetch_number_in_text(page_text, search_word = NULL)
}
\value{
\preformatted{ returns an integer
}
}
\description{
fetch a number from text
}
