% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_page}
\alias{get_page}
\title{\preformatted{  Retrieve an HTML object
}}
\usage{
get_page(url)
}
\value{
\preformatted{ an xml2 page object
}
}
\description{
simply a wrapper to retrieve an HTML page object from xml2
}
