% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{get_scraper_functions}
\alias{get_scraper_functions}
\title{\preformatted{  retrieve all the scraper functions
}}
\usage{
get_scraper_functions()
}
\value{
\preformatted{ a character containg the function to call
}
}
\description{
creates all state fxns and then add additional scrapes
}
