% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getDatabaseConnection}
\alias{getDatabaseConnection}
\title{\preformatted{  Establish a Database Connection
}}
\usage{
getDatabaseConnection(dbname = NULL, port = NULL, host = NULL,
  user = NULL, pass = NULL, retry = 5)
}
\arguments{
\item{dbname}{the name of the database to connect}

\item{port}{the port to connect to the database on}

\item{host}{the host name of the database}

\item{user}{the user to connect as}

\item{pass}{the password of the user to connect as}

\item{retry}{the number of times to retry before giving up}
}
\value{
\preformatted{ a database connection object
}
}
\description{
Using environemntal variables, establish a database connection
if no arguments are passed to it. Otherwise database connection
parameters can be established. The database driver used by this
function is \code{\link[RPostgres]{Postgres}()}.
}
