% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{getDbAsDt}
\alias{getDbAsDt}
\title{\preformatted{  retrieve a database table as a data.table
}}
\usage{
getDbAsDt(con, tbl, schema, dt = TRUE)
}
\arguments{
\item{con}{the database connection object}

\item{tbl}{the name of the table}

\item{schema}{the name of the schema}

\item{dt}{bool indicating whether to return a data.table or data.frame}
}
\value{
\preformatted{ either a \code{data.table} or \code{data.frame}
}
}
\description{
sends a query to retrieve the specified schema.table from a
database connection
}
