% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{lat_lon_to_fips}
\alias{lat_lon_to_fips}
\title{\preformatted{  convert lat/lon to fips
}}
\usage{
lat_lon_to_fips(points_df)
}
\value{
\preformatted{ returns the county_names found
}
}
\description{
provided a points_df containg lat/lon columns to
}
