% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{make_dat}
\alias{make_dat}
\title{\preformatted{  make the standardized data format
}}
\usage{
make_dat(country = "US", state = NA_character_, url = NA_character_,
  page = NA_character_, resolution = NA_character_,
  cases = NA_integer_, update_date = NA, deaths = NA,
  presumptive_positive = NA_integer_, recovered = NA_integer_,
  tested = NA_integer_, hospitalized = NA_integer_,
  county = NA_character_, negative_tests = NA_integer_,
  counties = NA_character_, severe_cases = NA_integer_,
  lat = NA_real_, no_longer_monitored = NA_integer_, lon = NA_real_,
  parish = NA_character_, fips = NA_integer_,
  monitored = NA_integer_, pending_tests = NA_integer_,
  active = NA_integer_, inconclusive = NA_integer_,
  quarantined = NA_integer_)
}
\value{
\preformatted{ returns a data.table ready for export or other usage
}
}
\description{
this provides the standard data format for import into the
covid database
}
