% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{make_datetime_utc}
\alias{make_datetime_utc}
\title{\preformatted{  convert a date to UTC
}}
\usage{
make_datetime_utc(x)
}
\value{
\preformatted{ a datetime object in UTC
}
}
\description{
converts a datetime object to UTC timezone
}
