% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{make_file_names}
\alias{make_file_names}
\title{\preformatted{  make file names
}}
\usage{
make_file_names(fxn)
}
\value{
\preformatted{  a list of file_names for csv data and html/json text
}
}
\description{
makes file names based on
}
