% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{retrieve_timezone}
\alias{retrieve_timezone}
\title{\preformatted{  lookup a posix-compatible timezone
}}
\usage{
retrieve_timezone(ab = NULL, dst = NULL, country = "US")
}
\value{
\preformatted{ the valid posix timezone
}
}
\description{
uses a combination of timezone datasets to make something
useful
}
