% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{retrieve_timzone_lut}
\alias{retrieve_timzone_lut}
\title{\preformatted{  retrieve the timzone lookup table from wikipedia
}}
\usage{
retrieve_timzone_lut()
}
\value{
\preformatted{ a data.table ready for insertion into a database or other use
}
}
\description{
this function is used to create the dataset to lookup
}
