% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{run_all_scripts}
\alias{run_all_scripts}
\title{\preformatted{  run all scripts
}}
\usage{
run_all_scripts()
}
\value{
\preformatted{ a data.table containing all data
}
}
\description{
runs all scripts from get_scraper_functions
}
