% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{sanitize_char}
\alias{sanitize_char}
\title{\preformatted{  sanitize a character
}}
\usage{
sanitize_char(x)
}
\value{
\preformatted{ a character
}
}
\description{
unlist and then create to character
}
