% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{sanitize_integer}
\alias{sanitize_integer}
\title{\preformatted{  sanitize an integer
}}
\usage{
sanitize_integer(x)
}
\value{
\preformatted{ returns an integer
}
}
\description{
convert whatever to an integer
}
