% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{sanitize_numeric}
\alias{sanitize_numeric}
\title{\preformatted{  sanitize a number
}}
\usage{
sanitize_numeric(x)
}
\value{
\preformatted{ returns a numeric number
}
}
\description{
given a possible list of numbers, create a numeric value
}
