% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_alabama}
\alias{scrape_alabama}
\title{\preformatted{  Scrape the Alabama Website
}}
\usage{
scrape_alabama()
}
\value{
\preformatted{ a data.table ready for use or import into database
}
}
\description{
Scrape the Alabama public health website for information about
coronavirus
}
