% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_arizona}
\alias{scrape_arizona}
\title{\preformatted{   Scrape the Arizona URLs
}}
\usage{
scrape_arizona()
}
\value{
\preformatted{ a data.table ready for use or import into database
}
}
\description{
Alaska has two URLs, one with the total cases, this script
parses both URLs and returns a single data.table
}
