% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_california}
\alias{scrape_california}
\title{\preformatted{ scrape la times from california
}}
\usage{
scrape_california()
}
\value{
\preformatted{ a data.table ready for use or import into database
}
}
\description{
scrapes data from the LA Times from the Case by county table
}
