% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_connecticut}
\alias{scrape_connecticut}
\title{\preformatted{ scrape connecticut
}}
\usage{
scrape_connecticut(driver = NULL)
}
\value{
\preformatted{ a data.table ready for use or import into database
}
}
\description{
scrapes urls for connecticut
}
