% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_delaware}
\alias{scrape_delaware}
\title{\preformatted{ scrape delaware
}}
\usage{
scrape_delaware()
}
\value{
\preformatted{ a data.table ready for use or import into database
}
}
\description{
scrapes urls for delaware
}
