% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_idaho}
\alias{scrape_idaho}
\title{\preformatted{ scrape DC
}}
\usage{
scrape_idaho()
}
\value{
\preformatted{ a data.table ready for use or import into database
}
}
\description{
scrapes urls for DC
}
