% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_new_hampshire}
\alias{scrape_new_hampshire}
\title{\preformatted{ scrape DC
}}
\usage{
scrape_new_hampshire()
}
\value{
\preformatted{ a data.table ready for use or import into database
}
}
\description{
scrapes urls for DC
}
