% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{scrape_puerto_rico}
\alias{scrape_puerto_rico}
\title{\preformatted{ scrape DC
}}
\usage{
scrape_puerto_rico()
}
\value{
\preformatted{ a data.table ready for use or import into database
}
}
\description{
scrapes urls for DC
}
