% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapers.R
\name{script_new_mexico}
\alias{script_new_mexico}
\title{\preformatted{ scrape DC
}}
\usage{
script_new_mexico()
}
\value{
\preformatted{ a data.table ready for use or import into database
}
}
\description{
scrapes urls for DC
}
