% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{unlist_coordinates}
\alias{unlist_coordinates}
\title{\preformatted{  unlist coordinates to points
}}
\usage{
unlist_coordinates(x)
}
\value{
\preformatted{ returns a data.table with lat/lon values
}
}
\description{
unlist a list of coordinates to points
}
