% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{wrap_scraper}
\alias{wrap_scraper}
\title{\preformatted{  scraper wrapper for scraper functions
}}
\usage{
wrap_scraper(fxn, interactive = FALSE)
}
\value{
\preformatted{ if interactive returns the data.table otherwise returns NULL
}
}
\description{
wrap the scraper in a
}
