% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{writeTable}
\alias{writeTable}
\title{\preformatted{ wrapper function around RPostgres::dbWriteTable
}}
\usage{
writeTable(x, con, tbl, schema, ...)
}
\description{
\preformatted{ wrapper function around RPostgres::dbWriteTable
}
}
